+++ config/api_router.py
@@ -1,4 +1,5 @@
 from django.conf import settings
+from django.urls import include, path
 from rest_framework.routers import DefaultRouter, SimpleRouter
 
 from kkanbu.users.api.views import UserViewSet
@@ -12,4 +13,9 @@ router.register("users", UserViewSet)
 
 
 app_name = "api"
-urlpatterns = router.urls
+
+urlpatterns = [
+    path("board/", include("kkanbu.board.urls")),
+]
+
+urlpatterns += router.urls